pose-service/requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0
scipy==1.11.2
scikit-image==0.21.0
tqdm==4.66.1
